refactors endpoint to MVC, including a return of the created username and id to the model
