+++ api/auth.py
@@ -11,9 +11,11 @@ from api.models.user_models import User
 from api.schemas.user_schemas import CreateUserRequest
 from api.schemas.authentication_schemas import Token
 
+from api.crud.auth_crud import create_user
+
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from jose import jwt
-import bcrypt
+
 
 router = APIRouter(
     prefix='/auth',
@@ -30,30 +32,6 @@ db_dependency = Annotated[Session, Depends(get_db)]
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
-def create_user(db: db_dependency, create_user_request: CreateUserRequest):
-
-    existing_user = db.query(User).filter(
-        User.username == create_user_request.username).first()
-    if existing_user:
-        raise HTTPException(status_code=409, detail="Username already exists.")
-
-    create_user_model = User(
-        username=create_user_request.username,
-        hashed_password=hash_password(create_user_request.password)
-    )
-
-    db.add(create_user_model)
-    db.commit()
-    db.refresh(create_user_model)
-
-
-def hash_password(password):
-    pwd_bytes = password.encode('utf-8')
-    salt = bcrypt.gensalt()
-    hashed_password = bcrypt.hashpw(password=pwd_bytes, salt=salt)
-    return hashed_password
-
-
-def verify_password(plain_password, hashed_password):
-    password_byte_enc = plain_password.encode('utf-8')
-    return bcrypt.checkpw(password=password_byte_enc, hashed_password=hashed_password)
+def post_user(db: db_dependency, create_user_request: CreateUserRequest):
+    user = create_user(db=db, request=create_user_request)
+    return {"message": "User created successfully, please log in to continue."}
